rename um/km registry helpers
