+++ libs/store_helper/kernel/ebpf_registry_helper.c
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation
 // SPDX-License-Identifier: MIT
 
-#include "ebpf_registry_helper_km.h"
+#include "ebpf_registry_helper.h"
 
 #define _EBPF_RESULT(x) (NT_SUCCESS(x) ? EBPF_SUCCESS : EBPF_FAILED)
