+++ libs/store_helper/user/CMakeLists.txt
@@ -2,14 +2,15 @@
 # SPDX-License-Identifier: MIT
 
 add_library("ebpf_store_helper_um" STATIC
-  ebpf_registry_helper_um.cpp
-  ebpf_store_helper.cpp
+  ebpf_registry_helper.cpp
+  ebpf_store_helper.c
 )
 
 target_include_directories("ebpf_store_helper_um" PRIVATE
   "${CMAKE_SOURCE_DIR}/include"
   "${CMAKE_SOURCE_DIR}/libs/platform"
   "${CMAKE_SOURCE_DIR}/libs/platform/user"
+  "${CMAKE_SOURCE_DIR}/libs/store_helper"
   "${CMAKE_SOURCE_DIR}/libs/store_helper/user"
   "${CMAKE_SOURCE_DIR}/external/usersim/inc"
   "${CMAKE_SOURCE_DIR}/external/ebpf-verifier/src"
