+++ libs/store_helper/user/ebpf_registry_helper.cpp
